fix comment on assimilating a 0
The note was based on the previous implementation, which shifted the
entire accumulator for each new bit.  The new implementation, introduced
in version 016 (using the manual version tracking system), adds a bit to
the start of the accumulator so no shifting of the accumulator is
required and the unfilled part of the accumulator (at the front of the
accumulator) always contains 0s.