multi-track awareness, no malloc, fields change
The main change here is that bc_decode now processes all tracks at once
instead of processing each track individually.  The library now uses a
struct bc_input and struct bc_decoded to specify the input and output
data.  All parts of the library and driver work with this change.  There
are still some minor issues, including track awareness in the field
list, but they will be fixed soon.

The other main, but less important, change is the removal of all
dynamic memory allocation.  This was done after analyzing the memory
footprint tradeoff of static allocation, which turned out to be around
20 kilobytes.  Since that size will only be an issue for very tiny
embedded devices, static allocation was chosen.  This change also allows
the library user to take advantage of the buffers created by the library
instead of allocating its own.

A minor change included in this commit is a change in the storing of
field data.  Instead of using one array that contains field names in the
even indices and field values in the odd indices, this version uses
separate arrays for the names and values.  This simplifies the code a
fair bit and makes it easier to read.  The change was made easier by the
decision to use static allocation (no need to optimize on the number of
malloc calls).