empty fields list after each card spec is read
This change fixes a bug that caused the "Type of card" field to appear
with the name of the last card in the formats file if the passed-in card
did not match any of the formats.  With this change, the library will
return an empty fields list if a card does not match any of the cards in
the formats file.