change name from special field to struct element
Make name an element in struct bc_decoded instead of a specially-handled
field in the fields list.  This is much cleaner.

Fixed a minor bug that may have caused garbage to be returned if the
formats file were empty.  Now initializing the name and fields list
before starting to process the formats file.

The last commit (9dfe947) introduced a minor bug fix to correct the
value of rc if bc_decode_track_fields on track 3 returns 0.  The fix
failed to make the commit log so it is being mentioned now.