add support for encoding type in formats file
The formats file must now contain the encoding type for each track.
This can be ALPHA, BCD, binary, unknown or none.  The last two refer to
cards where the track was not read or the card does not have any data on
that track, respectively.

In addition to the PCRE, libbitconvert now matches on the encoding type
as well.  So a track will only match if both the encoding type and the
PCRE for that track match the data on the track.

Included in this change is a new method of cleanup in
bc_decode_track_fields, which uses a goto and labels.  This simplifies
the cleanup process in the case of an error and ensures that the pointer
in the file handle points to the next track when the function returns.

BC_ENCODING_NONE was added in this change to represent tracks that do
not contain any bits.  These are different from tracks that contain
entirely binary 0s (which should be labelled BC_ENCODING_BINARY).  This
change also includes a shortcut in bc_decode that sets BC_ENCODING_NONE
when appropriate, avoiding a call to bc_decode_format.