expose the encoding type through the driver
Previously the driver did not print the encoding type, even though that
information was returned in struct bc_decoded.  Now the encoding is
printed along with the rest of the information in struct bc_decoded.

This changes adds the encoding_to_str function to the driver.  This
function should probably be in the library.  I will add it to the
library once I determine what the standard way is to do constant to
string conversions.