move field decoding into bc_find_fields
The field decoding part of the library has been broken out of bc_decode
and put in bc_find_fields instead.  In the case of an error, this makes
it easier for a user of the library to see where that error occurred:
during the bit decoding stage or the field decoding (finding) stage.