switch BCERR_* and BCINT_* to positive values
Convert the BCERR_* and BCINT_* values (previously all negative) to
positive values.  Negative return codes are unnecessary since the return
value is used exclusively for reporting errors (the conversion results
are stored entirely in structs, not passed through the return value, so
we don't need to keep the positive values free).