rename to_character function to to_ascii
The name of the to_character function did not accurately reflect what
the function does: convert bits into ASCII.  It was confusing because
the input and the output were both of the char (character) type.  As a
result, the name has been changed to to_ascii based on a suggestion from
Stephen Paul Weber.