add bc_combine for combining swipes from same card
bc_combine combines the data from forward and backward swipes of a
particular card.  Currently only track 2 is combined for testing
purposes.

A driver for bc_combine was added (combine.c).  It accepts a set of
forward swipes in the first 3 lines of standard in, followed by a set of
backward swipes in the second 3 lines of standard in.

BCERR_UNIMPLEMENTED was added to the error codes for referring to
unimplemented behavior in bc_combine.

Testing the bc_combine driver on 16 different cards showed that the
swipes from 13 cards could be successfully combined, while the swipes of
3 cards could not.  This should improve with future changes to
bc_combine.