memory for struct bc_input now managed by user
Instead of being statically-allocated in the library, the memory for
struct bc_input is now managed by the user of the library.  The test
drivers have been updated to reflect this; they each allocate their own
static buffers which are passed into the struct bc_input.

As a side-effect, bc_combine now allocates memory since it is the owner
of the struct bc_input "combined".  The user of bc_combine must thus
free the output using the new function bc_input_free.

Because of the change, bc_init no longer accepts a struct bc_input
because no initialization is required anymore.  This also eliminates the
meaning ambiguity with the function name -- does it mean initialize a
struct bc_input or initialize the library?  It now means initialize the
library.