add dynamic_fgets, which reads into growing buffer
The dynamic_fgets function is functionally-equivalent to fgets except it
grows the passed-in buffer as necessary and continues copying the line
contents into the buffer until a newline is reached.

Updated bc_decode_track_fields to use dynamic_fgets for the track
description line, which contains the encoding type and regular
expression.  This is read into the newly-created buf2.  Eventually all
fgets operations will be migrated to dynamic_fgets.