dynamic_fgets: change interface, fix several bugs
dynamic_fgets cannot use the same interface as fgets because
dynamic_fgets has more error conditions than fgets and because the
location of the buffer may change when it is reallocated.  Previously,
the extra error condition was ignored and the buffer location was
assumed not to change, which worked in small test cases but would
probably fail eventually.

The dynamic_fgets interface has been updated to return an int for error
conditions and to take a char** for the buffer so that a new buffer
location (possible after realloc) can be returned to the caller.  For
now, the error conditions are out of memory and EOF found.

dynamic_fgets has also been changed to take an int* instead of an int
for the buffer size.  This is so that the caller knows what the buffer's
new size is.  This is not strictly necessary (the caller doesn't need to
know this), but it is useful in that it allows multiple calls to
dynamic_fgets using the same buffer.  Reusing the buffer allows for
fewer memory allocation and free operations.