simplify field name finding code
Implement suggestions from Dave Vandervies to make the field name
finding code simpler and more readable.  As a side effect, two local
variables were eliminated.

The only functional change here is that a space between the period and
field name is now enforced.  Since this was never relied-on in the
formats file, no changes are needed there.  Aside from that, the new
code appears to be functionally-equivalent to the old code.