replace last use of fgets with dynamic_fgets
This change replaces the last of use fgets (which was in
bc_decode_track_fields) with dynamic_fgets.  As a side effect, the
FORMAT_LEN constant is no longer needed because the library accepts
lines in the format file of any length due to the use of dynamic_fgets.

Instead of using buf and buf2, which had been added for the first use of
dynamic_fgets, bc_decode_track_fields now uses only buf since the buffer
is handled by dynamic_fgets throughout the function.  As a result, the
buffer's space can be reused, which reduces the number of allocations
and frees required.

Because of the dynamically-sized buffer, strcpy calls have been changed
to strncpy for copying into field_names and field_values to prevent
overflowing the destination buffers, which have not yet been updated to
be dynamically-sized.