dynamically allocate name field in result
In bc_decode_fields, the result name is now set by assigning a pointer
instead of strncpy.  This change also moves that assignment into the if
block where a match has been found in order to prevent unnecessary
operations.

Added initialization of name in bc_decode to ensure name has a value set
even if the user doesn't run bc_find_fields.  This is necessary so that
bc_decoded_free will work in that case.