make ovector dynamically allocated (the last one)
ovector, the vector passed into pcre_exec that holds the offsets of the
captured substrings, is now dynamically allocated.  It is grown until
pcre_exec returns non-zero, indicating that ovector was large enough or
an error occurred.

This change eliminates MAX_CAPTURED_SUBSTRINGS, which is no longer
needed now that ovector is dynamically-sized.

This change is the last one needed to make libbitconvert completely free
of static buffers.  The test drivers (driver and combine) still have
static buffers, but the library itself does not.