move formats file to formats.txt
The formats file is now located at "formats.txt" instead of "formats".
This change was made to make it easier for Windows users to modify the
formats file.  Double-clicking on the formats file in Windows will now
open it in Notepad.  Developers on Windows should turn on the
"core.autocrlf" option in git:

	git config --global core.autocrlf true

This will cause the checked out files to use Windows-style newlines, which
Notepad can read.  Without this option, the newlines will be Unix-style
and Notepad will not read them correctly.

Although the formats file could have been stored in the repository with
Windows-style newlines, that option is not feasible because the
libbitconvert codebase cannot handle such a formats file when compiled on
a Unix-like system.  However, the libbitconvert codebase works with such a
formats file when compiled with the MinGW compiler on Windows.  As a
result, storing the formats file with Unix-style newlines and having git
deal with the newline conversion on Windows is the most feasible option.