use $(AR), $(RM), and local variables
Replace "ar" and "rm" with "$(AR)" and "$(RM)" so that the Makefile uses
the correct ar and rm.  Note that "$(RM)" expands to "rm -f" on GNU and
Solaris systems (probably POSIX) so an explicit "-f" is unneeded.

Replace explicit target and source with "$@" and "$<" to make the archive
rule more versatile (easier to copy into new rules).

These changes were suggested by Stephen Paul Weber.