add support for custom static libpcre
When building, the Makefile now checks if ../pcre exists and uses it for
the libpcre include and library files.  It assumes that such a libpcre
should be linked statically with libbitconvert.  This option is useful on
systems where a binary distributable that uses libbitconvert should contain
both libbitconvert and libpcre, such as on Windows and Mac OS X.