add "Bit Inspector test bitstream" to formats file
Add the Bit Inspector test bitstream, which is displayed in the
instructions shown at the top of the Bit Inspector window, to the formats
file.  This allows users to see how the fields table in Bit Inspector
works using the default test bitstream.  Previously, it may not have been
obvious to the user that a fields table existed unless they found another
bitstream to use that was recognized by the formats file.