bc_strerror: put case and return on separate lines
In bc_strerror, move the return statement to the line below the case
statement for each case.  These are purely whitespace changes; there is no
functional change.  This change is being made in preparation for the next
change, which will add to the existing error strings.  It is preferable to
keep these as separate changes so that it's easier to see which error
strings are added to.